lemur-broker/requirements.txt
Flask==2.0.1
google-cloud-pubsub==2.7.0
google-cloud-logging==2.6.0
